depositing for customer done

+++ SpringJava/zapp/src/main/java/com/app/zapp/service/CustomerService.java
@@ -18,7 +18,7 @@ public interface CustomerService {
 	
 	public String  getPassword(String id);
 	
-	public void depositAmount(String id,int amount);
+	public void depositAmount(String id,float amount);
 	
 	//public void editCustomer(Customer customer);
 }
